# swagger.yaml
openapi: 3.0.0
info:
  title: Student Management API
  description: A RESTful API for managing student records
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /:
    get:
      summary: Welcome endpoint
      description: Returns API information and links to documentation
      responses:
        '200':
          description: Welcome message with API information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Welcome'
  /students:
    get:
      summary: Get all students
      description: Retrieves a list of all students in the database
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Student'
        '500':
          description: Internal server error
    
    post:
      summary: Create a new student
      description: Add a new student to the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '201':
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid input
        '500':
          description: Internal server error

  /students/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the student
        schema:
          type: integer
          format: int64
    
    get:
      summary: Get a student by ID
      description: Retrieves a specific student by their ID
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '404':
          description: Student not found
        '500':
          description: Internal server error
    
    put:
      summary: Update a student
      description: Update an existing student's information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentInput'
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Student'
        '400':
          description: Invalid input
        '404':
          description: Student not found
        '500':
          description: Internal server error
    
    delete:
      summary: Delete a student
      description: Remove a student from the database
      responses:
        '204':
          description: Student deleted successfully
        '404':
          description: Student not found
        '500':
          description: Internal server error

components:
  schemas:
    Student:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: Unique identifier for the student
        name:
          type: string
          description: Full name of the student
        email:
          type: string
          format: email
          description: Email address of the student
        linkedin_profile:
          type: string
          description: URL to the student's LinkedIn profile
        phone:
          type: string
          description: Contact phone number of the student
      required:
        - id
        - name
        - email
    
    StudentInput:
      type: object
      properties:
        name:
          type: string
          description: Full name of the student
        email:
          type: string
          format: email
          description: Email address of the student
        linkedin_profile:
          type: string
          description: URL to the student's LinkedIn profile
        phone:
          type: string
          description: Contact phone number of the student
      required:
        - name
        - email

    Welcome:
      type: object
      properties:
        message:
          type: string
          description: Welcome message
          example: "Welcome to Students API"
        description:
          type: string
          description: Brief description of the API
          example: "A RESTful API for managing student records with CRUD operations"
        swagger_url:
          type: string
          description: URL to the Swagger documentation
          example: "/swagger"
        version:
          type: string
          description: API version
          example: "1.0.0"